# docs/swagger.json
{
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This API provides an access to data relevant to a championship.",
    "title": "Championship package.",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
...
    "/championships/{championshipID}/seasons": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "Championship"
        ],
        "summary": "Get filtered ChampionshipsSeasons.",
        "operationId": "getChampionshipsSeasons",
        "parameters": [
          {
            "type": "string",
            "name": "championshipID",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "seasonID",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GetChampionshipsSeasonsResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
    "/championships/{championshipID}/seasons/{seasonID}": {
      "post": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "Championship"
        ],
        "summary": "Register Season for Championship.",
        "operationId": "RegisterSeason",
        "parameters": [
          {
            "type": "string",
            "name": "championshipID",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "seasonID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/RegisterSeasonResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
...
    "/clubs": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "Club"
        ],
        "summary": "Get a list of clubs.",
        "operationId": "getClubs",
        "responses": {
          "200": {
            "$ref": "#/responses/GetClubsResponse"
          }
        }
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Club"
        ],
        "summary": "Create a new Club.",
        "operationId": "CreateClub",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/PostClubRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/PostClubResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
    "/clubs/{id}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Club"
        ],
        "summary": "Get a single club by ID.",
        "operationId": "getClub",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GetClubResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      },
      "put": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Club"
        ],
        "summary": "Update existing club.",
        "operationId": "UpdateClub",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/PutClubRequest"
            }
          },
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "PutClubResponse",
            "schema": {
              "$ref": "#/definitions/PutClubResponse"
            }
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      },
      "delete": {
        "tags": [
          "Club"
        ],
        "summary": "Delete a club.",
        "operationId": "DeleteClub",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
    "/seasons": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "Season"
        ],
        "summary": "Get a list of seasons.",
        "operationId": "getSeasons",
        "responses": {
          "200": {
            "$ref": "#/responses/GetSeasonsResponse"
          }
        }
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Season"
        ],
        "summary": "Create a new Season.",
        "operationId": "CreateSeason",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/PostSeasonRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/PostSeasonResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
    "/seasons/{id}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "Season"
        ],
        "summary": "Get a single Season by ID.",
        "operationId": "getSeason",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GetSeasonResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      },
      "delete": {
        "tags": [
          "Season"
        ],
        "summary": "Delete a season.",
        "operationId": "DeleteSeason",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    }
  },
  "definitions": {
    "ChampionshipDTO": {
      "type": "object",
      "required": [
        "id",
        "name"
      ],
      "properties": {
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-package": "_/home/ivan/Work/golang/api/rest/model"
    },
    "ChampionshipSeasonDTO": {
      "type": "object",
      "properties": {
        "Championship": {
          "$ref": "#/definitions/ChampionshipDTO"
        },
        "Season": {
          "$ref": "#/definitions/SeasonDTO"
        }
      },
      "x-go-package": "_/home/ivan/Work/golang/api/rest/model"
    },
...
    "ChampionshipsSeasonsDTO": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/ChampionshipSeasonDTO"
      },
      "x-go-package": "_/home/ivan/Work/golang/api/rest/model"
    },
    "ClubDTO": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-package": "_/home/ivan/Work/golang/api/rest/model"
    },
    "ClubsDTO": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/ClubDTO"
      },
      "x-go-package": "_/home/ivan/Work/golang/api/rest/model"
    },
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-package": "_/home/ivan/Work/golang/api/rest/model"
    },
...
    "GetChampionshipsSeasonsResponse": {
      "type": "object",
      "properties": {
        "ChampionshipsSeasons": {
          "$ref": "#/definitions/ChampionshipsSeasonsDTO"
        }
      },
      "x-go-package": "_/home/ivan/Work/golang/api/rest/model"
    },
    "GetClubResponse": {
      "type": "object",
      "properties": {
        "club": {
          "$ref": "#/definitions/ClubDTO"
        }
      },
      "x-go-package": "_/home/ivan/Work/golang/api/rest/model"
    },
    "GetClubsResponse": {
      "type": "object",
      "properties": {
        "clubs": {
          "$ref": "#/definitions/ClubsDTO"
        }
      },
      "x-go-package": "_/home/ivan/Work/golang/api/rest/model"
    },
    "GetSeasonResponse": {
      "type": "object",
      "properties": {
        "season": {
          "$ref": "#/definitions/SeasonDTO"
        }
      },
      "x-go-package": "_/home/ivan/Work/golang/api/rest/model"
    },
    "GetSeasonsResponse": {
      "type": "object",
      "properties": {
        "seasons": {
          "$ref": "#/definitions/SeasonsDTO"
        }
      },
      "x-go-package": "_/home/ivan/Work/golang/api/rest/model"
    },
...
    "PostClubRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-package": "_/home/ivan/Work/golang/api/rest/model"
    },
...
    "PostSeasonRequest": {
      "type": "object",
      "properties": {
        "endYear": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "EndYear"
        },
        "startYear": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StartYear"
        }
      },
      "x-go-package": "_/home/ivan/Work/golang/api/rest/model"
    },
    "PostSeasonResponse": {
      "type": "object",
      "properties": {
        "Season": {
          "$ref": "#/definitions/SeasonDTO"
        }
      },
      "x-go-package": "_/home/ivan/Work/golang/api/rest/model"
    },
...
    "PutClubRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-package": "_/home/ivan/Work/golang/api/rest/model"
    },
    "PutClubResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-package": "_/home/ivan/Work/golang/api/rest/model"
    },
    "RegisterSeasonResponse": {
      "type": "object",
      "properties": {
        "ChampionshipSeason": {
          "$ref": "#/definitions/ChampionshipSeasonDTO"
        }
      },
      "x-go-package": "_/home/ivan/Work/golang/api/rest/model"
    },
    "SeasonDTO": {
      "type": "object",
      "properties": {
        "EndYear": {
          "type": "integer",
          "format": "int64"
        },
        "ID": {
          "type": "string"
        },
        "StartYear": {
          "type": "integer",
          "format": "int64"
        }
      },
      "x-go-package": "_/home/ivan/Work/golang/api/rest/model"
    },
    "SeasonsDTO": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/SeasonDTO"
      },
      "x-go-package": "_/home/ivan/Work/golang/api/rest/model"
    }
  },
  "responses": {
    "ErrorResponse": {
      "description": "Standard error reponse.",
      "schema": {
        "$ref": "#/definitions/ErrorResponse"
      }
    },
...
    "GetChampionshipsSeasonsResponse": {
      "description": "Get ChampionshipsSeasons response",
      "schema": {
        "$ref": "#/definitions/GetChampionshipsSeasonsResponse"
      }
    },
    "GetClubResponse": {
      "description": "Single Club response",
      "schema": {
        "$ref": "#/definitions/GetClubResponse"
      }
    },
    "GetClubsResponse": {
      "description": "List of Clubs response.",
      "schema": {
        "$ref": "#/definitions/GetClubsResponse"
      }
    },
    "GetSeasonResponse": {
      "description": "Single Season response",
      "schema": {
        "$ref": "#/definitions/GetSeasonResponse"
      }
    },
    "GetSeasonsResponse": {
      "description": "List of Seasons response.",
      "schema": {
        "$ref": "#/definitions/GetSeasonsResponse"
      }
    },
...
    "PostClubResponse": {
      "description": "Update Club response",
      "schema": {
        "$ref": "#/definitions/PutClubResponse"
      }
    },
    "PostSeasonResponse": {
      "description": "Create Season response",
      "schema": {
        "$ref": "#/definitions/PostSeasonResponse"
      }
    },
    "RegisterSeasonResponse": {
      "description": "Create ChampionshipSeason response",
      "schema": {
        "$ref": "#/definitions/RegisterSeasonResponse"
      }
    }
  }
}